OOP implemented
